backend/pyproject.toml
[project]
name = "paper-reading-agent"
version = "0.1.0"
description = "AI-powered academic paper reading and analysis agent"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["ai", "paper-reading", "gemini", "pdf", "academic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "google-genai>=0.2.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "pypdf2>=3.0.0",
    "werkzeug>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "providers", "models"]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
